backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=43.0.0
httpx>=0.27.0
apscheduler>=3.10.4
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
